[refactor]: 펀딩취소모달 위치 변경(#118)
다음과 같은이유로 펀딩취소모달의 위치를 변경함

- 모달은 어떤 페이지에 종속되는(레이아웃) 개념 아님.
- 지금은 한 페이지에서만 사용되는 모달일지라도, 다른 곳에서 재사용될 수 있음.
